DAO ja controller luokat fiksattu (yritys ja asiakas).

+++ siri/src/main/java/fi/javaee/siri/asiakas/AsiakasDAO.java
@@ -58,28 +58,26 @@ public class AsiakasDAO implements Serializable {
 		return asiakas;
 	}
 
-	public void delete(Long id) {
-		@SuppressWarnings("unchecked")
-		List<Asiakas> asiakkaat = (List<Asiakas>) entityManager
-				.createQuery("select t from Asiakas t where t.asiakasId=:id").setParameter("id", id).getResultList();
-		for (Asiakas asiakas : asiakkaat) {
+	public Asiakas delete(Long id) {
+		Asiakas asiakas = findById(id);
+		
+		if (asiakas != null) {
 			entityManager.remove(asiakas);
-			break;
 		}
+
+		return asiakas;
 	}
 
-	// TODO edit -> fingById
-	public Asiakas edit(Long id) {
-		Asiakas tempAsiakas = null;
-		@SuppressWarnings("unchecked")
-		List<Asiakas> asiakkaat = (List<Asiakas>) entityManager
-				.createQuery("select t from Asiakas t where t.asiakasId=:id").setParameter("id", id).getResultList();
-		for (Asiakas asiakas : asiakkaat) {
-			tempAsiakas = asiakas;
-			break;
+	public Asiakas findById(Long id) {
+		Asiakas asiakas = null;
+		
+		try {
+			asiakas = (Asiakas) entityManager.createQuery("select t from Asiakas t where t.asiakasId=:id").setParameter("id", id).getSingleResult();
+		} catch (Exception e) {
+		    System.err.println("Caught Exception: " + e.getMessage());
 		}
-
-		return tempAsiakas;
+		
+		return asiakas;
 	}
 
 }
